split public run function for native and docker
