+++ fv3config/fv3run/__init__.py
@@ -1,1 +1,2 @@
-from ._run import run
+from ._docker import run_docker
+from ._native import run_native
