+++ fv3config/fv3run/__main__.py
@@ -1,6 +1,7 @@
 import sys
 import argparse
-from ._run import run
+from ._docker import run_docker
+from ._native import run_native
 
 MODULE_NAME = 'fv3config.run'
 STDOUT_FILENAME = 'stdout.log'
@@ -29,6 +30,33 @@ def main():
     """
     args = _parse_args()
     run(args.config, args.outdir, args.runfile, args.dockerimage, args.keyfile)
+
+
+def run(config_dict_or_location, outdir, runfile=None, docker_image=None, keyfile=None):
+    """Run the FV3GFS model with the given configuration.
+
+    Copies the resulting directory to a target location. Will use the Google cloud
+    storage key at ``$GOOGLE_APPLICATION_CREDENTIALS`` by default. Requires the
+    fv3gfs-python package.
+
+    Args:
+        config_dict_or_location (dict or str): a configuration dictionary, or a
+            location (local or on Google cloud storage) of a yaml file containing
+            a configuration dictionary
+        outdir (str): location to copy the resulting run directory
+        runfile (str, optional): Python model script to use in place of the default.
+        docker_image (str, optional): If given, run this command inside a container
+            using this docker image. Image must have this package and fv3gfs-python
+            installed.
+        keyfile (str, optional): location of a Google cloud storage key
+    """
+    if docker_image is not None:
+        run_docker(
+            config_dict_or_location, outdir, docker_image,
+            runfile=runfile, keyfile=keyfile
+        )
+    else:
+        run_native(config_dict_or_location, outdir, runfile=runfile)
 
 
 if __name__ == '__main__':
